requirement.txt
click==6.7
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Dropzone==1.5.3
Flask-Login==0.4.1
Flask-PyMongo==2.1.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gevent==1.3.6
greenlet==0.4.15
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycrypto==2.6.1
pymongo==3.7.1
visitor==0.1.3
